pyproject.toml
[project]
name = "pn-aoai-aisearch"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-search-documents (==11.5.3,<12.0.0)",
    "openai (>=1.93.3,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "azure-identity (>=1.23.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
